moved IArtboard to its server
