+++ app/definitions/artboard-generator.ts
@@ -1,3 +1,4 @@
+import { type IArtboard } from '#app/models/artboard.server'
 import { type IFill } from '#app/models/fill.server'
 import { type ILayer } from '#app/models/layer.server'
 import { type ILayout } from '#app/models/layout.server'
@@ -7,7 +8,6 @@ import { type IRotate } from '#app/models/rotate.server'
 import { type ISize } from '#app/models/size.server'
 import { type IStroke } from '#app/models/stroke.server'
 import { type ITemplate } from '#app/models/template.server'
-import { type IArtboard } from '#app/utils/db.server'
 
 // object sent to the client
 // generator has final building blocks for the generation(s)
