+++ app/routes/sketch+/artboards+/$slug_+/components/forms/design/panel-form-design-delete.tsx
@@ -3,10 +3,10 @@ import { getFieldsetConstraint } from '@conform-to/zod'
 import { useActionData, useFetcher } from '@remix-run/react'
 import { AuthenticityTokenInput } from 'remix-utils/csrf/react'
 import { PanelIconButton } from '#app/components/ui/panel-icon-button'
+import { type IArtboard } from '#app/models/artboard.server'
 import { type IDesignIdOrNull, type IDesign } from '#app/models/design.server'
 import { type ILayer } from '#app/models/layer.server'
 import { type DeleteDesignSchemaType } from '#app/schema/design'
-import { type IArtboard } from '#app/utils/db.server'
 import { useIsPending } from '#app/utils/misc'
 import { type IntentDesignDelete } from '../../../intent'
 import { type action } from '../../../route'
