+++ app/routes/sketch+/artboards+/$slug_+/components/popovers/layer/panel-popover-layer.tsx
@@ -1,8 +1,8 @@
 import { SidebarPanelPopoverFormContainer } from '#app/components/templates'
 import { Label } from '#app/components/ui/label'
 import { Separator } from '#app/components/ui/separator'
+import { type IArtboard } from '#app/models/artboard.server'
 import { type ILayer } from '#app/models/layer.server'
-import { type IArtboard } from '#app/utils/db.server'
 import { PanelFormArtboardLayerDelete } from '../../forms/artboard/layer/panel-form-artboard-layer-delete'
 import { PanelFormLayerEditDescription } from '../../forms/layer/panel-form-layer-edit-description'
 import { PanelFormLayerEditName } from '../../forms/layer/panel-form-layer-edit-name'
