+++ app/routes/sketch+/artboards+/$slug_+/services/artboard/design/delete.service.ts
@@ -1,4 +1,5 @@
 import { type User } from '@prisma/client'
+import { type IArtboard } from '#app/models/artboard.server'
 import {
 	type IDesign,
 	deleteDesign,
@@ -8,7 +9,7 @@ import {
 	connectPrevAndNextDesigns,
 } from '#app/models/design.server'
 import { type designTypeEnum } from '#app/schema/design'
-import { prisma, type IArtboard } from '#app/utils/db.server'
+import { prisma } from '#app/utils/db.server'
 import { artboardUpdateSelectedDesignService } from './update-selected.service'
 
 export const artboardDesignDeleteService = async ({
