+++ app/routes/sketch+/artboards+/$slug_+/services/artboard/layer/move-up.service.ts
@@ -1,11 +1,12 @@
 import { type User } from '@prisma/client'
+import { type IArtboard } from '#app/models/artboard.server'
 import {
 	findFirstLayer,
 	updateLayerRemoveNodes,
 	type ILayer,
 	updateLayerNodes,
 } from '#app/models/layer.server'
-import { type IArtboard, prisma } from '#app/utils/db.server'
+import { prisma } from '#app/utils/db.server'
 
 export const artboardLayerMoveUpService = async ({
 	userId,
